requirements.txt
azure.iot.device
python-dotenv==1.1.1